+++ backend/services/deck.service.js
@@ -1,8 +1,11 @@
-const { getDecksByCourseId } = require('../controllers/deck.controller');
 const deckModel = require('../models/deck.model');
+const courseModel = require('../models/course.model');
+const sequelize = require('sequelize');
+const azureStorage = require('../utils/azure-storage.util');
+const { DEFAULT_DECK_IMAGE } = require('../config/app.config');
 
 const deckService = {
-  getAllDecks: async (page = 1, limit = 10) => {
+  getAllDecks: async (page = 1, limit = 12) => {
     try {
       const offset = (page - 1) * limit;
       const { count, rows } = await deckModel.findAndCountAll({
@@ -15,6 +18,15 @@ const deckService = {
         totalPages: Math.ceil(count / limit),
         currentPage: page,
       };
+    } catch (error) {
+      console.log(error);
+      return error;
+    }
+  },
+  getDeckById: async (deckId) => {
+    try {
+      const deck = await deckModel.findByPk(deckId);
+      return deck;
     } catch (error) {
       console.log(error);
       return error;
@@ -59,6 +71,9 @@ const deckService = {
           image
         );
       }
+      if (imageUrl === null) {
+        imageUrl = DEFAULT_DECK_IMAGE;
+      }
 
       await deckModel.create({
         name,
@@ -66,14 +81,29 @@ const deckService = {
         image_url: imageUrl,
         course_id,
       });
+      await courseModel.update(
+        {
+          deck_count: sequelize.literal('deck_count + 1'),
+        },
+        {
+          where: {
+            id: course_id,
+          },
+        }
+      );
     } catch (error) {
       console.log(error);
       return error;
     }
   },
 
-  updateDeck: async ({ id, name, description, image }) => {
+  updateDeck: async ({ id, name, description, image, is_public }) => {
     try {
+      const deck = await deckModel.findByPk(id);
+      if (!deck) {
+        return null;
+      }
+      console.log('updateDeck', id, name, description, image, is_public);
       let imageUrl = null;
       if (image) {
         imageUrl = await azureStorage.uploadImage(
@@ -81,12 +111,15 @@ const deckService = {
           image
         );
       }
-
+      if (imageUrl === null) {
+        imageUrl = imageUrl || deck.image_url;
+      }
       await deckModel.update(
         {
           name,
           description,
           image_url: imageUrl,
+          is_public,
         },
         {
           where: {
@@ -99,26 +132,52 @@ const deckService = {
       return error;
     }
   },
-  deleteDeck: async (id) => {
+  deleteDeck: async (deck_id) => {
     try {
-      await deckModel.destroy({
+      const deck = await deckModel.findByPk(deck_id);
+      if (!deck) {
+        return null;
+      }
+      const course_id = deck.course_id;
+
+      const deletedDeck = await deckModel.destroy({
         where: {
-          id: id,
+          id: deck_id,
         },
       });
+
+      await courseModel.update(
+        {
+          deck_count: sequelize.literal('deck_count - 1'),
+        },
+        {
+          where: {
+            id: course_id,
+          },
+        }
+      );
+      return deletedDeck;
     } catch (error) {
       console.log(error);
       return error;
     }
   },
-  getDecksByCourseId: async (courseId) => {
+  getDecksByCourseId: async (courseId, page = 1, limit = 12) => {
     try {
-      const decks = await deckModel.findAll({
+      const offset = (page - 1) * limit;
+      const { count, rows } = await deckModel.findAndCountAll({
         where: {
           course_id: courseId,
         },
+        limit: limit,
+        offset: offset,
       });
-      return decks;
+      return {
+        total: count,
+        decks: rows,
+        totalPages: Math.ceil(count / limit),
+        currentPage: page,
+      };
     } catch (error) {
       console.log(error);
       return error;
